requirements.txt
Django==1.6.5
Fabric==1.9.0
Markdown==2.4.1
MySQL-python==1.2.5
Pillow==2.5.0
South==1.0
django-appconf==0.6
django-ckeditor-updated==4.4.0
django-compressor==1.4
django-constance==0.6
django-extensions==1.3.8
django-flatblocks==0.8
django-picklefield==0.3.1
django-registration==1.0
ecdsa==0.11
oauthlib==0.6.3
paramiko==1.14.0
pyScss==1.2.0.post3
pycrypto==2.6.1
python-openid==2.2.5
python-social-auth==0.1.26
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
sorl-thumbnail==11.12
wsgiref==0.1.2
